refactor: reorganize authentication forms and enhance component structure
- Updated imports in authentication forms to utilize a new centralized components structure, improving code organization.
- Introduced new forms for email verification, password reset, and forgot password functionalities, enhancing user experience.
- Refactored existing forms to use the new centralized components, ensuring consistency and clarity across the authentication flow.
- Improved error handling and user feedback mechanisms in forms for better usability.
